Resolves merge conflicts and co-locates definition wrt build tag
